fix: feature-loyalty-widget-and-score-refactor (#7791)
