+++ backend/plugins/loyalty_api/src/modules/score/db/models/ScoreLog.ts
@@ -194,15 +194,11 @@ export const loadScoreLogClass = (models: IModels, subdomain: string) => {
     }
 
     public static async getScoreLogs(doc: IScoreLogParams) {
-      const { stageId, pipelineId, boardId, number, orderType } = doc;
-      const limit = Math.min(Math.max(Number(doc.limit) || 50, 1), 100);
+      const { stageId, pipelineId, boardId, number } = doc;
+      const limit = Math.min(Math.max(Number(doc.limit) || 50, 1), 200);
       const direction = doc.direction === 'backward' ? 'backward' : 'forward';
-      const logsPerOwner = Math.min(
-        Math.max(Number(doc.logsPerOwner) || 5, 1),
-        100,
-      );
-      const orderBy: Record<string, SortOrder> =
-        orderType === 'createdAt' ? { createdAt: -1 } : { totalScore: -1 };
+
+      const orderBy: Record<string, SortOrder> = { createdAt: -1 };
       const sortFields = Object.keys(orderBy);
       const sortOrder = {
         ...Object.fromEntries(
@@ -238,24 +234,10 @@ export const loadScoreLogClass = (models: IModels, subdomain: string) => {
 
       const basePipeline: any[] = [
         ...filterAggregate,
-        {
-          $match: { ...filter },
-        },
-        {
-          $sort: { createdAt: -1, _id: -1 },
-        },
+        { $match: { ...filter } },
         {
           $addFields: {
             signedScore: buildSignedScoreExpression(),
-          },
-        },
-        {
-          $group: {
-            _id: '$ownerId',
-            ownerType: { $first: '$ownerType' },
-            logs: { $push: '$$ROOT' },
-            createdAt: { $max: '$createdAt' },
-            totalScore: { $sum: '$signedScore' },
           },
         },
       ];
@@ -263,31 +245,48 @@ export const loadScoreLogClass = (models: IModels, subdomain: string) => {
       const cursorMatch = doc.cursor
         ? buildCursorQuery(doc.cursor, orderBy, direction, {
             createdAt: 'date',
-            totalScore: 'number',
           })
         : null;
 
       const listPipeline: any[] = [
         ...basePipeline,
-        {
-          $project: {
-            _id: '$_id',
-            ownerId: '$_id',
-            ownerType: 1,
-            logs: { $slice: ['$logs', logsPerOwner] },
-            createdAt: 1,
-            totalScore: 1,
-          },
-        },
         ...(cursorMatch ? [{ $match: cursorMatch }] : []),
         { $sort: sortOrder },
         { $limit: limit + 1 },
+        {
+          $lookup: {
+            from: 'score_logs',
+            let: { ownerId: '$ownerId' },
+            pipeline: [
+              { $match: { $expr: { $eq: ['$ownerId', '$$ownerId'] } } },
+              {
+                $addFields: { signedScore: buildSignedScoreExpression() },
+              },
+              {
+                $group: {
+                  _id: '$ownerId',
+                  totalScore: { $sum: '$signedScore' },
+                },
+              },
+            ],
+            as: '_ownerTotal',
+          },
+        },
+        {
+          $addFields: {
+            totalScore: {
+              $ifNull: [{ $arrayElemAt: ['$_ownerTotal.totalScore', 0] }, 0],
+            },
+          },
+        },
+        { $project: { _ownerTotal: 0, signedScore: 0 } },
       ];
 
       const [listRaw, countResult] = await Promise.all([
         models.ScoreLogs.aggregate(listPipeline).allowDiskUse(true),
         models.ScoreLogs.aggregate([
-          ...basePipeline,
+          ...filterAggregate,
+          { $match: { ...filter } },
           { $count: 'totalCount' },
         ]).allowDiskUse(true),
       ]);
@@ -298,6 +297,31 @@ export const loadScoreLogClass = (models: IModels, subdomain: string) => {
       if (direction === 'backward') {
         list = list.reverse();
       }
+
+      const ownerKeys = Array.from(
+        new Set(
+          list
+            .filter((l: any) => l.ownerId && l.ownerType)
+            .map((l: any) => `${l.ownerType}:${l.ownerId}`),
+        ),
+      );
+
+      const ownerMap = new Map<string, any>();
+      await Promise.all(
+        ownerKeys.map(async (key) => {
+          const [ownerType, ownerId] = key.split(':');
+          const owner = await getLoyaltyOwner(subdomain, {
+            ownerType,
+            ownerId,
+          });
+          if (owner) ownerMap.set(key, owner);
+        }),
+      );
+
+      list = list.map((item: any) => ({
+        ...item,
+        owner: ownerMap.get(`${item.ownerType}:${item.ownerId}`) || null,
+      }));
 
       const pageInfo: PageInfo = {
         hasNextPage: direction === 'forward' ? hasMore : Boolean(doc.cursor),
