+++ backend/plugins/loyalty_api/src/modules/score/graphql/schemas/scoreLog.ts
@@ -25,12 +25,28 @@ export const types = `
   }
 
   type ScoreLog {
+    _id: String
     ownerId: String
     ownerType: String
-
     owner: JSON
-    logs: [ScoreLogItem]
     totalScore: Float
+
+    change: Float
+    action: String
+    description: String
+
+    campaignId: String
+    campaign: ScoreCampaign
+
+    targetId: String
+    target: JSON
+
+    serviceName: String
+    createdBy: String
+    createdAt: Date
+
+    amount: Float
+    quantity: Float
   }
 
   type ScoreLogList {
