+++ frontend/plugins/loyalty_ui/src/modules/loyalties/score/components/ScoreRecordTable.tsx
@@ -5,76 +5,62 @@ import {
   SCORE_LOG_CURSOR_SESSION_KEY,
   useScoreList,
 } from '../hooks/useScoreList';
-import { IScoreLog } from '../types/score';
 import { GiveScoreModal } from './GiveScoreModal';
-import { ScoreDetailSheet } from './ScoreDetailSheet';
-import { useState, useMemo } from 'react';
+import { useMemo } from 'react';
 
 export const ScoreRecordTable = () => {
   const { list, loading, handleFetchMore, pageInfo } = useScoreList();
   const { hasPreviousPage, hasNextPage } = pageInfo || {};
-  const [selectedRecord, setSelectedRecord] = useState<IScoreLog | null>(null);
 
-  const columns = useMemo(
-    () => makeScoreColumns((row) => setSelectedRecord(row)),
-    [],
-  );
+  const columns = useMemo(() => makeScoreColumns(), []);
 
   if (loading && !list?.length) return <Spinner />;
 
   return (
-    <>
-      <RecordTable.Provider
-        columns={columns}
-        data={list || []}
-        className="m-3 relative"
-        stickyColumns={['more', 'ownerName']}
+    <RecordTable.Provider
+      columns={columns}
+      data={list || []}
+      className="m-3 relative"
+      stickyColumns={['more', 'ownerName']}
+    >
+      <RecordTable.CursorProvider
+        hasPreviousPage={hasPreviousPage}
+        hasNextPage={hasNextPage}
+        dataLength={list?.length}
+        sessionKey={SCORE_LOG_CURSOR_SESSION_KEY}
       >
-        <RecordTable.CursorProvider
-          hasPreviousPage={hasPreviousPage}
-          hasNextPage={hasNextPage}
-          dataLength={list?.length}
-          sessionKey={SCORE_LOG_CURSOR_SESSION_KEY}
-        >
-          <RecordTable>
-            <RecordTable.Header />
-            <RecordTable.Body>
-              <RecordTable.CursorBackwardSkeleton
-                handleFetchMore={handleFetchMore}
-              />
-              {loading ? (
-                <RecordTable.RowSkeleton rows={32} />
-              ) : (
-                <RecordTable.RowList />
-              )}
-              <RecordTable.CursorForwardSkeleton
-                handleFetchMore={handleFetchMore}
-              />
-            </RecordTable.Body>
-          </RecordTable>
+        <RecordTable>
+          <RecordTable.Header />
+          <RecordTable.Body>
+            <RecordTable.CursorBackwardSkeleton
+              handleFetchMore={handleFetchMore}
+            />
+            {loading ? (
+              <RecordTable.RowSkeleton rows={32} />
+            ) : (
+              <RecordTable.RowList />
+            )}
+            <RecordTable.CursorForwardSkeleton
+              handleFetchMore={handleFetchMore}
+            />
+          </RecordTable.Body>
+        </RecordTable>
 
-          {!loading && list?.length === 0 && (
-            <div className="absolute inset-0 flex items-center justify-center">
-              <div className="flex flex-col items-center text-center">
-                <IconStar size={48} className="text-gray-400 mb-4" />
-                <h3 className="text-lg font-semibold text-gray-900">
-                  No scores yet
-                </h3>
-                <p className="mt-1 text-sm text-gray-500 mb-4">
-                  Get started by giving your first score.
-                </p>
-                <GiveScoreModal />
-              </div>
+        {!loading && list?.length === 0 && (
+          <div className="absolute inset-0 flex items-center justify-center">
+            <div className="flex flex-col items-center text-center">
+              <IconStar size={48} className="text-gray-400 mb-4" />
+              <h3 className="text-lg font-semibold text-gray-900">
+                No scores yet
+              </h3>
+              <p className="mt-1 text-sm text-gray-500 mb-4">
+                Get started by giving your first score.
+              </p>
+              <GiveScoreModal />
             </div>
-          )}
-        </RecordTable.CursorProvider>
-      </RecordTable.Provider>
-
-      <ScoreDetailSheet
-        open={!!selectedRecord}
-        onOpenChange={(open) => !open && setSelectedRecord(null)}
-        record={selectedRecord}
-      />
-    </>
+          </div>
+        )}
+      </RecordTable.CursorProvider>
+    </RecordTable.Provider>
   );
 };
